[pymc3] Add suffix to model variables: avoid clobbering other variables
My code, as well as PyMC3, assumes all variables have a unique name. So if a
model variable and an associated distribution share the same name, this can
cause one to shadow the other (or just throw an error down the line).
This is avoided by adding a customizable suffix to distribution variable names.
By default the suffix is " (dist)".