[parameters] Allow consistent filenames with different Numpy versions
With v1.14, Numpy updating its string representation algorithm for arrays, with
altered the computed filenames. This commit adds support for fully specifying
the print options in different Numpy versions (the new version has more
options), a mechanism for requesting 'legacy' printing (so v1.14 can open
filenames computed with earlier versions) and a warning if v1.13 is used with
this safeguard in place.

Commit also adds a test for this functionality in tests/parameters_test.py.