[utils] Add SanitizedOrderedDict
SanitizedOrderedDict subclasses OrderedDict, and adds the ability to sanitize
keys. E.g. if entries are keyed by display text, this may contain formatting
characters or codes codes which are difficult to retype from memory. This class
allows this keys to be 'sanitized', i.e. for a set of characters to be removed
before they are used to index the dictionary.
Alternatively, an arbitrary sanitization function can also be provided.
This is intended for user-facing dictionaries, to allow the user to access
entries without needing to remember the exact key, as long as the one they
specify is "close enough".