[theano] Rolling dependency cache.
Rather than immediately deleting a cache when dependencies change, it is 
put in a stack, which by default has a length of 5. When this length is 
exceeded, the least recently used cache is deleted. This avoids 2 
issues:
- If a model is used twice (e.g. to generate data and then to fit to 
it), the two graphs don't destroy each other's cache.
- When debugging, running with `pdb.set_trace()` somewhere does not 
destroy the 'real' cache.