[parameterized] Fix support for `Computed` parameters. Separate module
Fixed support for type annotations indicating an attribute is 
'computed'.

Moved the @parameterized decorator code to a separate module. Among 
other things, this avoids a name clash with `Any`, which is also used 
for parameter parsing / validation.
(In fact, @dataclass and @parameterized likely obviate the need for the 
parameter validation functions in `parameters`.)