Add json encoder for complex numbers. Make 'field' arguments optional
'field' args: Sometimes we wanted to use a validator on its own, outside 
a Pydantic model where there isn't necessarily a 'field'. Since the 
field argument is only used for the error message, giving it a default 
value (name="") is a good compromise to allow these methods to be used 
more widely.