[parameters:ParameterComparison] Remove progress bar when complete
The reason for the progress bar is that sometimes a parameter comparison 
takes a while, and we want to give some progress feedback to the user. 
But when there are multiple parameter sets to compare, we end up with a 
pile of useless progress bars at the end.
Now the bars are removed once the comparison is complete.